requirements.txt
langchain
langchain-google-genai
langchain-mcp-adapters
mcp
uvicorn
fastapi
psutil
python-dotenv